refactor: signup 로직 수정
