--- src/main/java/techcourse/myblog/service/DuplicatedEmailException.java
@@ -1,9 +1,0 @@
-package techcourse.myblog.service;
-
-public class DuplicatedEmailException extends RuntimeException {
-    public static final String DUPLICATED_USER_MESSAGE = "이미 존재하는 email입니다";
-
-    public DuplicatedEmailException() {
-        super(DUPLICATED_USER_MESSAGE);
-    }
-}
