+++ src/main/java/techcourse/myblog/service/LoginFailedException.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.web.controller;
+package techcourse.myblog.service;
 
 public class LoginFailedException extends RuntimeException {
     public static final String LOGIN_FAIL_MESSAGE = "이메일이나 비밀번호가 올바르지 않습니다";
