+++ src/main/java/techcourse/myblog/service/SignUpFailedException.java
@@ -1,0 +1,7 @@
+package techcourse.myblog.service;
+
+public class SignUpFailedException extends RuntimeException {
+    public SignUpFailedException(String msg) {
+        super(msg);
+    }
+}
